Use external libtiff by defining TIFF_LIB macro
Macros in nmake.opt have been defined to instruct the make process
to use an external version of libtiff.

Unfortunately, the makefiles assume these macros contain absolute
paths. As a workaround, the original macros are defined relative
to two levels deep from the root, e.g., relative to frmts\gtiff.
This works fine for subprojects that rely on these macros.
However, the main makefile uses TIFF_LIB in defining the
EXTERNAL_LIBS macro. For this to work, TIFF_LIB would need to be
relative from the project root. The workaround here is to
introduce a new TIFF_LIB_FROM_ROOT macro and use that one instead
in the definition of EXTERNAL_LIBS.