Include libtiff.lib in static lib output
When compiling GDAL with an external TIFF lib, the user is required
to specify this additional dependency when compiling a program
using GDAL. This is not the case when GDAL is compiled with the
internal TIFF functionality. By having the external libtiff.lib
included in the output gdal.lib, the user only has to link with
gdal.lib, simplifying the linking process for the user.